requirements.txt
discord.py>=2.2.2
discord.py[voice]==2.2.2
pyffmpeg==2.4.2.16
ffmpeg-python>=0.2.0
PyPDF2==3.0.1
Pillow==9.4.0
docx2pdf>=0.1.8
pdf2docx>=0.5.6
yt-dlp==2023.3.4
openai-whisper>=20230308